requirements.txt
black==20.8b1
click==7.1.2
click-log==0.3.2
flake8==3.8.4
flake8-black==0.2.1
flake8-import-order==0.18.1
mypy==0.812
Pillow==8.1.0
pytest==6.2.2
